delete redundant definitions
