+++ src/bench/comparison_v2.py
@@ -2,11 +2,12 @@ import csv
 import json
 import os
 import time
+from typing import List
 
 import httpx
 import pandas as pd
 
-from src.auto_acmg import AutoACMG, AutoACMGPrediction
+from src.auto_acmg import AutoACMG
 from src.core.config import settings
 from src.defs.genome_builds import GenomeRelease
 from src.defs.seqvar import SeqVar
@@ -147,10 +148,10 @@ def eval_autoacmg(pred, expected):
     :return: criteria met, true positives, false negatives, false positives
     :rtype: tuple
     """
-    crit_met = []
-    for crit in pred.model_dump().values():
-        if crit["prediction"] == AutoACMGPrediction.Met:
-            crit_met.append(crit["name"])
+    crit_met: List[str] = []
+    # for crit in pred.model_dump().values():
+    #     if crit["prediction"] == AutoACMGPrediction.Met:
+    #         crit_met.append(crit["name"])
     tp = list(set(expected) & set(crit_met))
     fn = list(set(expected) - set(crit_met))
     fp = list(set(crit_met) - set(expected))
