+++ src/defs/annonars_gene.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 
 from typing import Any, Dict, List, Optional
 
-from pydantic import BaseModel, Field, RootModel, validator
+from pydantic import BaseModel, RootModel
 
 
 class Clingen(BaseModel):
